On OpenBSD, default java.net.preferIPv4Stack to true and disable setting IPV6_V6ONLY socket option.
OpenBSD does not support IPv4-mapped IPv6 sockets at all. For Unix
like operating systems, the JDK relies on this feature to support
IPv4 and IPv6 with one socket. Consequently, the JDK on OpenBSD can
only support one protocol at a time. This change makes the default
be IPv4 which, even today, is the predominantly used protocol.

The proper solution is to make the JDK use dual sockets to support
both IPv6 and IPv4 like it does for windows.